+++ src/test/java/com/postech/tabletrust/repository/RestaurantRepositoryIT.java
@@ -1,6 +1,7 @@
 package com.postech.tabletrust.repository;
 
 import com.postech.tabletrust.entities.Restaurant;
+import com.postech.tabletrust.utils.NewEntititesHelper;
 import jakarta.transaction.Transactional;
 import org.hibernate.type.descriptor.java.LocalTimeJavaType;
 import org.junit.jupiter.api.Test;
@@ -39,7 +40,7 @@ public class RestaurantRepositoryIT {
     @Test
     void shouldAllowSaveRestaurant() {
         //Arrange
-        Restaurant restaurant = createARestaurant();
+        Restaurant restaurant = NewEntititesHelper.createARestaurant();
 
         //Act
         var restaurantFoud = restaurantRepository.save(restaurant);
@@ -54,36 +55,17 @@ public class RestaurantRepositoryIT {
     @Test
     void shouldFindRestaurantsByNameAndAddressAndKitchenType() {
         // Arrange - cria e persiste um restaurante de teste
-        Restaurant restaurant = new Restaurant();
-        restaurant.setName("Test Restaurant");
-        restaurant.setAddress("Test Address");
-        restaurant.setKitchenType("Test Kitchen");
-        restaurant.setAvailableCapacity(10);
+        Restaurant restaurant = NewEntititesHelper.createARestaurant();
 
         entityManager.persist(restaurant);
         entityManager.flush();
 
         // Act - busca restaurantes usando o método do repositório
         List<Restaurant> foundRestaurants = restaurantRepository.findRestaurantsByNameAndAddressAndKitchenType(
-                "Test Restaurant", "Test Address", "Test Kitchen");
+                "Restaurante-teste", "Fragonard", "Sopa");
 
         // Assert - verifica se o restaurante correto foi encontrado
         assertThat(foundRestaurants).hasSize(1);
         assertThat(foundRestaurants.get(0).getId()).isEqualTo(restaurant.getId());
-    }
-
-    private Restaurant createARestaurant(){
-        LocalTime open = new LocalTimeJavaType().fromString("19:00:00");
-        LocalTime close = new LocalTimeJavaType().fromString("23:30:00");
-
-        return Restaurant.builder()
-                //.id(UUID.randomUUID())
-                .address("Fragonard")
-                .kitchenType("Tapioca")
-                .name("Restaurante-teste")
-                .openingTime(open)
-                .closingTime(close)
-                .availableCapacity(100)
-                .build();
     }
 }
